requirements.txt
python-docx==1.1.2
PyPDF2==3.0.1
pytesseract==0.3.13
Pillow==10.4.0
regex==2024.11.6
